--- src/features/backoffice/reports/reports.backoffice.view.container.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import ReportsBackofficeContainer from './reports.backoffice.container';
-
-const ReportsBackofficeViewContainer = () => {
-  return (
-    <div>
-      <ReportsBackofficeContainer />
-    </div>
-  );
-};
-
-export default ReportsBackofficeViewContainer;
